HW3/source/requirement.txt
asttokens==2.0.8
autopep8==2.0.0
av==9.2.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
einops==0.6.0
entrypoints==0.4
executing==1.1.0
fairscale==0.4.13
fonttools==4.37.4
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
labml==0.4.161
labml-helpers==0.4.89
labml-nn==0.4.133
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
pyspark==3.0.1
python-dateutil==2.8.2
pytorch-gan-metrics==0.5.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.5.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.12.1+cu116
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
